+++ cmd/commands.go
@@ -157,6 +157,7 @@ var (
 		AddonsAddCommand,
 		AddonsRemoveCommand,
 		AddonsUpgradeCommand,
+		AddonsInfoCommand,
 
 		// Integration Link
 		integrationLinkShowCommand,